refactor(graph): extract the note-file-type indexers (#1624)
Third per-format split of indexers.ts (arch D1): the non-markdown note file
types (.ttl / .csv / .py), called by indexNote's early returns.

- New `graph/indexers/note-files.ts`: indexTurtleFile, indexCsvFile,
  indexPythonFile. Their identical note-metadata setup (type/title/filename/
  path/mtime/folder/project) is deduped into one `addFileNoteMetadata` helper.
- `ensureFolder` — the last shared low-level helper these needed — moves to the
  leaf `index-helpers.ts` (beside ensureTag), so the split stays acyclic; the
  note-indexing path imports it from there.
- indexNote imports the three file-type indexers from the new module. They're
  internal (not facade-exported), so no facade change.
  indexers.ts: 1358 → 1210 lines (1669 at the start of the D1 splits).

Behaviour-preserving — full graph suite (440) green; lint clean.

Remaining D1 follow-up: the tables cluster (note-CSVW indexTable + the
tables-panel indexCsvTable/indexMarkdownTable).
